requirements.txt
numpy==1.21.5
paho-mqtt==1.5.1
pyyaml==6.0.2
requests==2.32.2
opencv-python==4.10.0.84
pillow==10.4.0
